% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.rmd-package.R
\docType{package}
\name{addins.rmd-package}
\alias{addins.rmd}
\alias{addins.rmd-package}
\title{addins.rmd: RStudio Add-ins to Format R Markdown Documents}
\description{
Package `addins.rmd` contains RStudio add-ins to format R Markdown documents. The main purpose of this package is to help begginers to learn R Markwown syntax faster. As well as to create possibility to use keyboard shortcuts for the most common R Markdown syntax elements.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://gegznav.github.io/addins.rmd/}
  \item Report bugs at \url{https://github.com/GegznaV/addins.rmd/issues}
}

}
\author{
\strong{Maintainer}: Vilmantas Gegzna \email{GegznaV@gmail.com} [copyright holder]

}
\keyword{internal}
