% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_addins.rmd-package.R
\docType{package}
\name{addins.rmd}
\alias{addins.rmd}
\alias{addins.rmd-package}
\title{RStudio Add-ins for R Markdown}
\description{
\code{addins.rmd} provides a set of RStudio add-ins to format R Markdown documents.

\code{License:} MIT   \cr
\code{URL:} \url{https://github.com/GegznaV/addins.rmd} \cr
\code{Bug reports and suggestions:} \url{https://github.com/GegznaV/addins.rmd/issues} \cr
\code{Author:} Vilmantas Gegzna
}
\seealso{
Use the links below to learn more about RStudio add-ins
and especially how to use them in combination with user-defined
keyboard shortcuts:\enumerate{

\item{\href{https://rstudio.github.io/rstudioaddins/}{RStudio add-ins};}
\item{\href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{Keyboard shortcuts};}
\item{\href{https://support.rstudio.com/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts}{Customizing keyboard shortcuts.}}
}
}
