% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_formatting.R, R/rmd_symbols_as_is.R
\name{format_rmd}
\alias{format_rmd}
\alias{rmd_format_bold}
\alias{rmd_format_italics}
\alias{rmd_format_bold_italics}
\alias{rmd_format_bold2}
\alias{rmd_format_italics2}
\alias{rmd_code_inline}
\alias{rmd_code_inline_r}
\alias{rmd_code_inline_highlighted_as_r}
\alias{rmd_format_superscript}
\alias{rmd_format_subscript}
\alias{rmd_format_strikethrough}
\alias{rmd_horizontal_rule}
\alias{rmd_link_url}
\alias{rmd_insert_figure}
\alias{rmd_insert_figure_block}
\alias{rmd_symbol_asterisk}
\alias{rmd_symbol_backslash}
\alias{rmd_symbol_underscore}
\alias{rmd_symbol_dollar}
\alias{rmd_symbol_en_dash}
\alias{rmd_symbol_em_dash}
\title{Add-ins for basic R Markdown formatting.}
\usage{
rmd_format_bold(context = rs_get_context())

rmd_format_italics(context = rs_get_context())

rmd_format_bold_italics(context = rs_get_context())

rmd_format_bold2(context = rs_get_context())

rmd_format_italics2(context = rs_get_context())

rmd_code_inline(context = rs_get_context())

rmd_code_inline_r(context = rs_get_context())

rmd_code_inline_highlighted_as_r(context = rs_get_context())

rmd_format_superscript(context = rs_get_context())

rmd_format_subscript(context = rs_get_context())

rmd_format_strikethrough(context = rs_get_context())

rmd_horizontal_rule(context = rs_get_context())

rmd_link_url(context = rs_get_context())

rmd_insert_figure(context = rs_get_context())

rmd_insert_figure_block(context = rs_get_context())

rmd_symbol_asterisk(context = rs_get_context())

rmd_symbol_backslash(context = rs_get_context())

rmd_symbol_underscore(context = rs_get_context())

rmd_symbol_dollar(context = rs_get_context())

rmd_symbol_en_dash(context = rs_get_context())

rmd_symbol_em_dash(context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\description{
RStudio add-ins, which enclose selected text with symbols that have special
interpretation in R Markdown.
}
\seealso{
https://pandoc.org/MANUAL.html#backslash-escapes

Other R Markdown formatting add-ins: \code{\link{rmd_code_block}},
  \code{\link{rmd_equations}}, \code{\link{rmd_list}}

Other R Markdown formatting add-ins: \code{\link{rmd_code_block}},
  \code{\link{rmd_equations}}, \code{\link{rmd_list}}
}
\concept{R Markdown formatting add-ins}
