% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_insert_various.R
\name{insert_various_rmd}
\alias{insert_various_rmd}
\alias{rmd_icon_fa_download}
\alias{rmd_icon_fa_external_link}
\alias{rmd_icon_fa_youtube_square}
\alias{rmd_icon_fa_file_pdf}
\title{Insert various elements}
\usage{
rmd_icon_fa_download()

rmd_icon_fa_external_link()

rmd_icon_fa_youtube_square()

rmd_icon_fa_file_pdf()
}
\description{
Insert various elements
}
\seealso{
Other R Markdown formatting add-ins: \code{\link{format_rmd}},
  \code{\link{rmd_code_block}},
  \code{\link{rmd_equations}}, \code{\link{rmd_list}}
}
