% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_tables.R
\name{rmd_tables}
\alias{rmd_tables}
\alias{rmd_table_style_md}
\alias{rmd_table_style_md_alternative}
\alias{rmd_table_style_multiline}
\alias{rmd_table_style_multiline_no_colnames}
\alias{rmd_table_style_simple}
\alias{rmd_table_style_grid}
\title{Insert R Markdown tables}
\usage{
rmd_table_style_md(context = rs_get_context())

rmd_table_style_md_alternative(context = rs_get_context())

rmd_table_style_multiline(context = rs_get_context())

rmd_table_style_multiline_no_colnames(context = rs_get_context())

rmd_table_style_simple(context = rs_get_context())

rmd_table_style_grid(context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\description{
Insert examples of R Markdown tables.
}
