% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_code_block.R
\name{rmd_code_block}
\alias{rmd_code_block}
\alias{rmd_code_block_r}
\alias{rmd_code_block_r_split}
\title{Convert rows into the block of code.}
\usage{
rmd_code_block_r(context = rs_get_context())

rmd_code_block(context = rs_get_context())

rmd_code_block_r_split(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
RStudio add-in to insert selected lines into code block: \itemize{
   \item \code{rmd_code_block()} - verbatim code block;
   \item \code{rmd_code_block_r()} - R code block.
}.
\code{rmd_code_block_r_split()} - splits block of R code.


These functions internally use function
\code{\link[addin.tools]{rs_enclose_selected_rows_with}()} -
that adds lines above and below the selection.
}
\seealso{
Other R Markdown formatting add-ins: \code{\link{format_rmd}},
  \code{\link{rmd_equations}}, \code{\link{rmd_list}}
}
\concept{R Markdown formatting add-ins}
