% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_headings.R
\name{rmd_headings}
\alias{rmd_headings}
\alias{rmd_heading_1}
\alias{rmd_heading_2}
\alias{rmd_heading_3}
\alias{rmd_heading_4}
\alias{rmd_heading_5}
\alias{rmd_heading_6}
\alias{rmd_heading_1_title}
\alias{rmd_heading_2_subtitle}
\title{Format text as R Markdown headings.}
\usage{
rmd_heading_1(context = rs_get_context())

rmd_heading_2(context = rs_get_context())

rmd_heading_3(context = rs_get_context())

rmd_heading_4(context = rs_get_context())

rmd_heading_5(context = rs_get_context())

rmd_heading_6(context = rs_get_context())

rmd_heading_1_title(context = rs_get_context())

rmd_heading_2_subtitle(context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\description{
RStudio add-ins to format text as R Markdown headings.
}
