% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_headings.R
\name{rmd_headings}
\alias{rmd_headings}
\alias{rmd_heading_1}
\alias{rmd_heading_2}
\alias{rmd_heading_3}
\alias{rmd_heading_4}
\alias{rmd_heading_5}
\alias{rmd_heading_6}
\alias{rmd_heading_1_title}
\alias{rmd_heading_2_subtitle}
\alias{rmd_heading_remove}
\title{Format text as R Markdown headings.}
\usage{
rmd_heading_1(context = rs_get_context())

rmd_heading_2(context = rs_get_context())

rmd_heading_3(context = rs_get_context())

rmd_heading_4(context = rs_get_context())

rmd_heading_5(context = rs_get_context())

rmd_heading_6(context = rs_get_context())

rmd_heading_1_title(context = rs_get_context())

rmd_heading_2_subtitle(context = rs_get_context())

rmd_heading_remove(style = c("auto", "both", "hash", "underline"),
  context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}

\item{style}{("both", "hash", "underline", "auto") indicates which style of heading to remove: \itemize{
    \item "underline" - removes underline-style heading formatting only;
    \item "hash" - removes hash-style heading formatting only;
    \item "both" - removes both heading formatting styles;
    \item "auto" - tries to remove only one style: if finds underline style,
    removes it, otherwise removes hash-style.
}}
}
\description{
RStudio add-ins to format text as R Markdown headings and to remove this formatting.
}
\details{
Function \code{rmd_heading_remove()} removes heading formatting:
either hash-symbol-style headings (i.e., remove leading hash symbols \code{#}
and spaces), or
underline-style headings in the first selected row. \cr

The other functions, which name beginns with \code{rmd_heading_}, at first
apply \code{rmd_heading_remove()} to remove appropriate heading style and
then format the first selected row as a heading of an appropriate level.
}
