% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_tables.R
\name{rmd_tables}
\alias{rmd_tables}
\alias{rmd_table_style_md}
\alias{rmd_table_style_multiline}
\alias{rmd_table_style_simple}
\alias{rmd_table_style_simple_no_colnames}
\alias{rmd_table_style_grid}
\title{Insert R Markdown tables}
\usage{
rmd_table_style_md()

rmd_table_style_multiline()

rmd_table_style_simple()

rmd_table_style_simple_no_colnames()

rmd_table_style_grid()
}
\description{
Insert examples of R Markdown tables.
}
